+++ src/containers/App.tsx
@@ -1,7 +1,7 @@
 import classnames from 'classnames'
 import { Route, Navigate, Routes, useLocation, Outlet } from 'react-router-dom'
 
-// import Overview from '@containers/Overview'
+import Overview from '@containers/Overview'
 import Connections from '@containers/Connections'
 import ExternalControllerModal from '@containers/ExternalControllerDrawer'
 import Logs from '@containers/Logs'
@@ -22,7 +22,7 @@ export default function App () {
     const location = useLocation()
 
     const routes = [
-    // { path: '/', name: 'Overview', component: Overview, exact: true },
+        { path: '/overview', name: 'Overview', element: <Overview /> },
         { path: '/proxies', name: 'Proxies', element: <Proxies /> },
         { path: '/logs', name: 'Logs', element: <Logs /> },
         { path: '/rules', name: 'Rules', element: <Rules />, noMobile: true },
@@ -43,7 +43,7 @@ export default function App () {
     return (
         <Routes>
             <Route path="/" element={layout}>
-                <Route path="/" element={<Navigate to={{ pathname: '/proxies', search: location.search }} replace />} />
+                <Route path="/" element={<Navigate to={{ pathname: '/overview', search: location.search }} replace />} />
                 {
                     routes.map(
                         route => <Route path={route.path} key={route.path} element={route.element} />,